+++ workers/unity/Assets/Fps/Scripts/Editor/SnapshotGenerator/SnapshotMenu.cs
@@ -1,5 +1,6 @@
 using System.IO;
 using Improbable.Gdk.Core;
+using Improbable.Gdk.DeploymentManager;
 using UnityEditor;
 using UnityEngine;
 
@@ -10,24 +11,37 @@ namespace Fps
         public static readonly string DefaultSnapshotPath =
             Path.Combine(Application.dataPath, "../../../snapshots/default.snapshot");
 
-        [MenuItem("SpatialOS/Generate FPS Snapshot")]
-        private static void GenerateDefaultSnapshot()
-        {
-            var snapshot = new Snapshot();
+        public static readonly string CloudSnapshotPath =
+            Path.Combine(Application.dataPath, "../../../snapshots/cloud.snapshot");
 
+        private static void GenerateSnapshot(Snapshot snapshot)
+        {
             var spawner = FpsEntityTemplates.Spawner();
             snapshot.AddEntity(spawner);
-
-            var SimulatedPlayerCoordinatorTrigger = FpsEntityTemplates.SimulatedPlayerCoordinatorTrigger();
-            snapshot.AddEntity(SimulatedPlayerCoordinatorTrigger);
-
-            SaveSnapshot(snapshot);
         }
 
-        private static void SaveSnapshot(Snapshot snapshot)
+        [MenuItem("SpatialOS/Generate FPS Snapshot")]
+        private static void GenerateFpsSnapshot()
         {
-            snapshot.WriteToFile(DefaultSnapshotPath);
-            Debug.LogFormat("Successfully generated initial world snapshot at {0}", DefaultSnapshotPath);
+            var localSnapshot = new Snapshot();
+            var cloudSnapshot = new Snapshot();
+
+            GenerateSnapshot(localSnapshot);
+            GenerateSnapshot(cloudSnapshot);
+
+            // The local snapshot is identical to the cloud snapshot, but also includes a simulated player coordinator
+            // trigger.
+            var simulatedPlayerCoordinatorTrigger = FpsEntityTemplates.SimulatedPlayerCoordinatorTrigger();
+            localSnapshot.AddEntity(simulatedPlayerCoordinatorTrigger);
+
+            SaveSnapshot(DefaultSnapshotPath, localSnapshot);
+            SaveSnapshot(CloudSnapshotPath, cloudSnapshot);
+        }
+
+        private static void SaveSnapshot(string path, Snapshot snapshot)
+        {
+            snapshot.WriteToFile(path);
+            Debug.LogFormat("Successfully generated initial world snapshot at {0}", path);
         }
     }
 }