+++ workers/unity/Assets/Fps/Scripts/SetupLogic/SimulatedPlayerWorkerConnector.cs
@@ -1,18 +1,48 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Fps;
 using Improbable.Gdk.Core;
 using Improbable.Gdk.GameObjectCreation;
 using Improbable.Gdk.GameObjectRepresentation;
 using Improbable.Gdk.PlayerLifecycle;
+using Improbable.Worker.CInterop.Alpha;
+using UnityEngine;
 
 public class SimulatedPlayerWorkerConnector : DefaultWorkerConnector
 {
     private const string AuthPlayer = "Prefabs/SimulatedPlayer/SimulatedPlayer";
     private const string NonAuthPlayer = "Prefabs/SimulatedPlayer/SimulatedPlayerProxy";
 
-    private async void Start()
+    private ILogDispatcher simulatedCoordinatorLogDispatcher;
+    private bool connectToRemoteDeployment;
+
+    private string simulatedPlayerDevAuthTokenId;
+    private string simulatedPlayerTargetDeployment;
+
+    public async Task ConnectSimulatedPlayer(ILogDispatcher logDispatcher, string simulatedPlayerDevAuthTokenId,
+        string simulatedPlayerTargetDeployment)
     {
-        await Connect(WorkerUtils.SimulatedPlayer, new ForwardingDispatcher());
+        simulatedCoordinatorLogDispatcher = logDispatcher;
+
+        // If we're connecting simulated players to another deployment, we need to fetch locator tokens.
+        if (!string.IsNullOrEmpty(simulatedPlayerTargetDeployment))
+        {
+            if (string.IsNullOrEmpty(simulatedPlayerDevAuthTokenId))
+            {
+                logDispatcher.HandleLog(LogType.Error,
+                    new LogEvent("Failed to launch simulated player. No development authentication token specified."));
+                return;
+            }
+
+            this.simulatedPlayerDevAuthTokenId = simulatedPlayerDevAuthTokenId;
+            this.simulatedPlayerTargetDeployment = simulatedPlayerTargetDeployment;
+
+            connectToRemoteDeployment = true;
+        }
+
+        await Connect(WorkerUtils.UnityClient, new ForwardingDispatcher());
     }
 
     protected override void HandleWorkerConnectionEstablished()
@@ -29,6 +59,11 @@ public class SimulatedPlayerWorkerConnector : DefaultWorkerConnector
             new AdvancedEntityPipeline(Worker, AuthPlayer, NonAuthPlayer, fallback));
     }
 
+    protected override ConnectionService GetConnectionService()
+    {
+        return connectToRemoteDeployment ? ConnectionService.AlphaLocator : ConnectionService.Receptionist;
+    }
+
     protected override ReceptionistConfig GetReceptionistConfig(string workerType)
     {
         var config = base.GetReceptionistConfig(workerType);
@@ -37,5 +72,40 @@ public class SimulatedPlayerWorkerConnector : DefaultWorkerConnector
         config.WorkerId = $"{workerType}-{Guid.NewGuid()}";
 
         return config;
+    }
+
+    protected override AlphaLocatorConfig GetAlphaLocatorConfig(string workerType)
+    {
+        var pit = GetDevelopmentPlayerIdentityToken(simulatedPlayerDevAuthTokenId, "SimulatedPlayer", GetDisplayName());
+        var loginTokens = GetDevelopmentLoginTokens(workerType, pit);
+        var loginToken = SelectLoginToken(loginTokens);
+
+        return new AlphaLocatorConfig
+        {
+            LocatorHost = RuntimeConfigDefaults.LocatorHost,
+            LocatorParameters = new LocatorParameters
+            {
+                PlayerIdentity = new PlayerIdentityCredentials
+                {
+                    LoginToken = loginToken,
+                    PlayerIdentityToken = pit
+                }
+            }
+        };
+    }
+
+    protected override string SelectLoginToken(List<LoginTokenDetails> loginTokens)
+    {
+        var selectedLoginToken = loginTokens.FirstOrDefault(token => token.DeploymentName == simulatedPlayerTargetDeployment)
+            .LoginToken;
+
+        if (selectedLoginToken == null)
+        {
+            simulatedCoordinatorLogDispatcher.HandleLog(LogType.Error,
+                new LogEvent(
+                    "Failed to launch simulated player. Login token for target deployment was not found in response. Does that deployment have the `dev_auth` tag?"));
+        }
+
+        return selectedLoginToken;
     }
 }